fix rest api format
We should follow rest api format. Moving app-profile to be hasura since that reflects it's resource more clearly
