+++ web/api/hasura/delete-images/index.ts
