+++ web/api/hasura/get-app-review-images/graphql/getAppReviewImages.graphql
