+++ web/api/hasura/verify-app/graphql/getAppMetadata.graphql
